+++ src/main/java/dao/interfaces/ColorDao.java
@@ -6,9 +6,9 @@ import java.sql.SQLException;
 import java.util.Optional;
 
 public interface ColorDao {
-    Optional<Integer> addColor(String title) throws SQLException;
+    Optional<Long> addColor(String title) throws SQLException;
     Optional<Color> getByTitle(String title) throws SQLException;
     Color getOrElseAddAndGet(String title) throws  SQLException;
-    Optional<Color> getById(int id) throws SQLException;
+    Optional<Color> getById(long id) throws SQLException;
     boolean remove(Color color) throws SQLException;
 }
